Написал свою версию telegrambot, в настоящий момент ещё не включающей в себя scheduler, функции, связанные с изменением испытательного срока, сохранение отчётов о состоянии животных и тесты к сервисам

+++ Animals/src/main/java/sky/pro/Animals/controller/VolunteerController.java
@@ -55,7 +55,7 @@ public class VolunteerController {
                                                     @RequestParam String address,
                                                     @RequestParam Date birthday,
                                                     @RequestParam String passport,
-                                                    @RequestParam Integer chatId,
+                                                    @RequestParam Long chatId,
                                                     @RequestParam String workPosition) {
         infoService.checkInfo();
         Volunteer volunteer = new Volunteer(id, fio, address, birthday, passport, chatId, workPosition);
@@ -73,7 +73,7 @@ public class VolunteerController {
                                                    @RequestParam String address,
                                                    @RequestParam Date birthday,
                                                    @RequestParam String passport,
-                                                   @RequestParam Integer chatId,
+                                                   @RequestParam Long chatId,
                                                    @RequestParam String workPosition) {
         infoService.checkInfo();
         Volunteer volunteer = new Volunteer(id, fio, address, birthday, passport, chatId, workPosition);
